google maps feature added inside consultant company info screen yet not working completely

+++ lib/screens/consultant_screens/widgets/requestWidgets.dart
@@ -10,29 +10,104 @@ class PendingRequest extends StatefulWidget {
       {required this.name,
       required this.projectDataList,
       required this.engEmail,
+      required this.selectedValue,
       super.key});
   String name;
   List projectDataList;
   String engEmail;
+  String selectedValue;
   @override
   State<PendingRequest> createState() => _PendingRequestState();
 }
 
 class _PendingRequestState extends State<PendingRequest> {
+  confirmContrReq() async {
+    try {
+      print('1');
+      var activitiesSnapshot = await FirebaseFirestore.instance
+          .collection('contractorReq')
+          .where('projectId', isEqualTo: widget.projectDataList[3])
+          .get();
+
+      var list = activitiesSnapshot.docs.map((e) {
+        var update = FirebaseFirestore.instance
+            .collection('contractorReq')
+            .doc(e.id)
+            .update({'reqAccepted': true});
+
+        return update;
+      }).toList();
+
+      var updateLongQuery = await FirebaseFirestore.instance
+          .collection('contractor')
+          .doc(widget.engEmail)
+          .collection('projects')
+          .where('projectId', isEqualTo: widget.projectDataList[3])
+          .get();
+      var longList = updateLongQuery.docs.map((e) async {
+        var accept_req = await FirebaseFirestore.instance
+            .collection('contractor')
+            .doc(widget.engEmail)
+            .collection('projects')
+            .doc(e.id)
+            .update({'reqAccepted': true});
+        return accept_req;
+      }).toList();
+    } catch (e) {}
+  }
+
+  rejecContrReq() async {
+    var activitiesSnapshot = await FirebaseFirestore.instance
+        .collection('contractorReq')
+        .where('projectId', isEqualTo: widget.projectDataList[3])
+        .get();
+    var list = activitiesSnapshot.docs.map((e) async {
+      var update = await FirebaseFirestore.instance
+          .collection('contractorReq')
+          .doc(e.id)
+          .delete();
+      return update;
+    }).toList();
+    var updateLongQuery = await FirebaseFirestore.instance
+        .collection('contractor')
+        .doc(widget.engEmail)
+        .collection('projects')
+        .where('projectId', isEqualTo: widget.projectDataList[3])
+        .get();
+    var longList = updateLongQuery.docs.map((e) async {
+      var accept_req = await FirebaseFirestore.instance
+          .collection('contractor')
+          .doc(widget.engEmail)
+          .collection('projects')
+          .doc(e.id)
+          .delete();
+      return accept_req;
+    }).toList();
+    await FirebaseFirestore.instance
+        .collection('Projects')
+        .doc(widget.projectDataList[3])
+        .update({'isContrSelected': false});
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         appBar: AppBar(
           elevation: 0,
-          title: Text('Pending Requests',style: TextStyle(color:Colors.black),),
+          title: Text(
+            'Pending Requests',
+            style: TextStyle(color: Colors.black),
+          ),
           centerTitle: true,
         ),
         body: Column(
           children: [
             RequestBody(
-                name: widget.name,
-                projectDataList: widget.projectDataList,
-                engEmail: widget.engEmail),
+              name: widget.name,
+              projectDataList: widget.projectDataList,
+              engEmail: widget.engEmail,
+              role: widget.selectedValue,
+            ),
             SizedBox(
               height: 10,
             ),
@@ -47,36 +122,43 @@ class _PendingRequestState extends State<PendingRequest> {
                     textColor: Colors.black,
                     icon: Icons.cloud_done_rounded,
                     onTap: () async {
-                      var activitiesSnapshot = await FirebaseFirestore.instance
-                          .collection('engineers')
-                          .doc('${widget.engEmail}')
-                          .update({'reqAccepted': true});
+                      var activitiesSnapshot =
+                          await widget.selectedValue == 'Contractor'
+                              ? confirmContrReq()
+                              : FirebaseFirestore.instance
+                                  .collection('engineers')
+                                  .doc('${widget.engEmail}')
+                                  .update({'reqAccepted': true});
                       Navigator.pop(context);
                       setState(() {});
                       Get.snackbar('Request Accepted',
-                          'Engineer has been added to the project');
+                          '${widget.selectedValue} has been added to the project');
                     },
                   ),
                   SizedBox(
                     width: 10,
                   ),
                   MyButton(
-                    text: 'Reject',
+                    text: ' Reject   ',
                     bgColor: Colors.red,
                     textColor: Colors.black,
                     icon: Icons.cancel,
                     onTap: () async {
-                      var activitiesSnapshot = await FirebaseFirestore.instance
-                          .collection('engineers')
-                          .doc('${widget.engEmail}')
-                          .delete(
-                        //FieldPath(['reqAccepted']): FieldValue.delete(),
-                      );
-                       await FirebaseFirestore.instance
-                      .collection('Projects')
-                      .doc(widget.projectDataList[3])
-                      .update({'isSelected': false});
-
+                      if (widget.selectedValue == 'Contractor') {
+                        rejecContrReq();
+                      } else {
+                        var activitiesSnapshot = await FirebaseFirestore
+                            .instance
+                            .collection('engineers')
+                            .doc('${widget.engEmail}')
+                            .delete(
+                                //FieldPath(['reqAccepted']): FieldValue.delete(),
+                                );
+                        await FirebaseFirestore.instance
+                            .collection('Projects')
+                            .doc(widget.projectDataList[3])
+                            .update({'isSelected': false});
+                      }
                       Navigator.pop(context);
                       setState(() {});
                       Get.snackbar('Request Rejected', '');
@@ -95,21 +177,59 @@ class ApprovedRequest extends StatefulWidget {
       {required this.name,
       required this.projectDataList,
       required this.engEmail,
+      required this.selectedValue,
       super.key});
   String name;
   List projectDataList;
   String engEmail;
+  String selectedValue;
   @override
   State<ApprovedRequest> createState() => _ApprovedRequestState();
 }
 
 class _ApprovedRequestState extends State<ApprovedRequest> {
+  rejecContrReq() async {
+    var activitiesSnapshot = await FirebaseFirestore.instance
+        .collection('contractorReq')
+        .where('projectId', isEqualTo: widget.projectDataList[3])
+        .get();
+    var list = activitiesSnapshot.docs.map((e) async {
+      var update = await FirebaseFirestore.instance
+          .collection('contractorReq')
+          .doc(e.id)
+          .delete();
+      return update;
+    }).toList();
+    var updateLongQuery = await FirebaseFirestore.instance
+        .collection('contractor')
+        .doc(widget.engEmail)
+        .collection('projects')
+        .where('projectId', isEqualTo: widget.projectDataList[3])
+        .get();
+    var longList = updateLongQuery.docs.map((e) async {
+      var accept_req = await FirebaseFirestore.instance
+          .collection('contractor')
+          .doc(widget.engEmail)
+          .collection('projects')
+          .doc(e.id)
+          .delete();
+      return accept_req;
+    }).toList();
+    await FirebaseFirestore.instance
+        .collection('Projects')
+        .doc(widget.projectDataList[3])
+        .update({'isContrSelected': false});
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
         appBar: AppBar(
           elevation: 0,
-          title: Text('Approved Requests',style: TextStyle(color:Colors.black),),
+          title: Text(
+            'Approved Requests',
+            style: TextStyle(color: Colors.black),
+          ),
           centerTitle: true,
         ),
         body: Column(
@@ -119,6 +239,7 @@ class _ApprovedRequestState extends State<ApprovedRequest> {
               name: widget.name,
               projectDataList: widget.projectDataList,
               engEmail: widget.engEmail,
+              role: widget.selectedValue,
             ),
             SizedBox(
               height: 20,
@@ -131,18 +252,22 @@ class _ApprovedRequestState extends State<ApprovedRequest> {
                 textColor: Colors.black,
                 icon: Icons.delete,
                 onTap: () async {
-                  var activitiesSnapshot = await FirebaseFirestore.instance
-                      .collection('engineers')
-                      .doc('${widget.engEmail}')
-                      .delete();
-                  await FirebaseFirestore.instance
-                      .collection('Projects')
-                      .doc(widget.projectDataList[3])
-                      .update({'isSelected': false});
+                  if (widget.selectedValue == 'Contractor') {
+                    rejecContrReq();
+                  } else {
+                    var activitiesSnapshot = await FirebaseFirestore.instance
+                        .collection('engineers')
+                        .doc('${widget.engEmail}')
+                        .delete();
+                    await FirebaseFirestore.instance
+                        .collection('Projects')
+                        .doc(widget.projectDataList[3])
+                        .update({'isSelected': false});
+                  }
 
                   Navigator.pop(context);
                   setState(() {});
-                  Get.snackbar('Engineer Deleted', '');
+                  Get.snackbar('${widget.selectedValue} Deleted', '');
                 },
               ),
             ),
@@ -156,10 +281,12 @@ class RequestBody extends StatefulWidget {
       {required this.name,
       required this.projectDataList,
       required this.engEmail,
+      required this.role,
       super.key});
   String name;
   List projectDataList;
   String engEmail;
+  String role;
 
   @override
   State<RequestBody> createState() => _RequestBodyState();
@@ -182,14 +309,20 @@ class _RequestBodyState extends State<RequestBody> {
                 CircleAvatar(
                   backgroundColor: Colors.green,
                   radius: 30,
-                  child: Icon(Icons.person,color:Colors.black,),
+                  child: Icon(
+                    Icons.person,
+                    color: Colors.black,
+                  ),
                 ),
                 SizedBox(
                   width: 10,
                 ),
                 Text(
                   '${widget.name}',
-                  style: TextStyle(fontSize: 20, color:Colors.black,fontWeight: FontWeight.bold),
+                  style: TextStyle(
+                      fontSize: 20,
+                      color: Colors.black,
+                      fontWeight: FontWeight.bold),
                 ),
               ],
             ),
@@ -210,6 +343,7 @@ class _RequestBodyState extends State<RequestBody> {
           Padding(
             padding: const EdgeInsets.all(32.0),
             child: Container(
+              width: 270,
               decoration: BoxDecoration(
                 borderRadius: BorderRadius.circular(10),
                 color: Colors.grey,
@@ -250,7 +384,7 @@ class _RequestBodyState extends State<RequestBody> {
                             'Role:  ',
                           ),
                           Text(
-                            'Engineer',
+                            widget.role,
                           ),
                         ],
                       ),
